+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 01. :gift: Added exception `dev.orne.test.rnd.GenerationException`
 01. :gift: Added exception `dev.orne.test.rnd.UnsupportedValueTypeException`
+01. :gift: Added annotation `dev.orne.test.rnd.Priority`
 01. :gift: Added interface `dev.orne.test.rnd.Generator`
 01. :gift: Added class `dev.orne.test.rnd.AbstractGenerator`
 01. :gift: Added interface `dev.orne.test.rnd.TypedGenerator`